+++ js/web-components/sortable-table.js
@@ -14,11 +14,15 @@
 /* eslint spaced-comment: ["error", "always"] */
 /* eslint-disable no-console */
 
+import { usingWebComponentsPolyfill } from './utils.js';
 import { getSortObject, smartSort } from './utils-sort.js';
 
 class SortableTable extends HTMLTableElement {
     constructor() {
         super();
+        if (usingWebComponentsPolyfill()) {
+            return;
+        }
         this.setupTable();
     }
 
@@ -156,7 +160,7 @@ class SortableTable extends HTMLTableElement {
 
 window.customElements.define('sortable-table', SortableTable, { extends: 'table' });
 
-// For Safari, Samsung Internet, and Edge
+// For Safari, Samsung Internet, and Legacy Edge
 window._webComponentPolyfills = window._webComponentPolyfills || [];
 window._webComponentPolyfills.push({
     element: 'sortable-table',